Huge update

+++ src/main/java/com/sjkz1/sjkz1misc/render/RainbowLayer.java
@@ -14,9 +14,6 @@ import net.minecraft.client.render.entity.feature.FeatureRendererContext;
 import net.minecraft.client.render.entity.model.EntityModel;
 import net.minecraft.client.util.math.MatrixStack;
 import net.minecraft.entity.Entity;
-import net.minecraft.entity.boss.dragon.EnderDragonEntity;
-import net.minecraft.entity.mob.EndermanEntity;
-import net.minecraft.util.math.MathHelper;
 
 public class RainbowLayer<T extends Entity, M extends EntityModel<T>> extends FeatureRenderer<T, M>
 {
@@ -30,7 +27,7 @@ public class RainbowLayer<T extends Entity, M extends EntityModel<T>> extends Fe
 
 	@Override
 	public void render(MatrixStack matrixStack, VertexConsumerProvider vertexConsumers, int light, T entity, float limbAngle, float limbDistance, float tickDelta, float animationProgress, float headYaw, float headPitch) {
-		float ticks = ((float)(entity.age % 20) + tickDelta) / 20.0F;
+		float ticks = (entity.age % 20 + tickDelta) / 20.0F;
 		Color color = Color.getHSBColor(ticks,0.9f,1);
 
 		if (!entity.isInvisible() && SJKZ1Misc.dance && !(entity instanceof AbstractClientPlayerEntity)) {
